server/requirements.txt
flask
pymongo
flask-pymongo
grpcio
grpcio-tools
python-decouple

numpy
# torch
# torchvision
python-chess